backend/requirements.txt
# iLyra Platform - Backend Dependencies
# Python 3.11+ required

# Core Flask Framework
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0

# Database
SQLAlchemy==2.0.23
PyMySQL==1.1.0
mysql-connector-python==8.2.0

# Security & Authentication
bcrypt==4.1.2
PyJWT==2.8.0
cryptography

# API & Serialization
marshmallow==3.20.2
Flask-Marshmallow==0.15.0
marshmallow-sqlalchemy==0.29.0

# Validation
email-validator==2.1.0
WTForms==3.1.1
Flask-WTF==1.2.1

# HTTP & Requests
requests==2.31.0
urllib3==2.1.0

# Date & Time
python-dateutil==2.8.2
pytz==2023.3

# Environment & Configuration
python-dotenv==1.0.0
python-decouple==3.8

# Development & Testing
pytest==7.4.3
pytest-flask==1.3.0
coverage==7.3.2

# Production Server
gunicorn==21.2.0
waitress==2.1.2

# Utilities
click==8.1.7
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
Werkzeug==3.0.1

# JSON & Data Processing
jsonschema==4.20.0
simplejson==3.19.2

# Logging & Monitoring
structlog==23.2.0

# Rate Limiting
Flask-Limiter==3.5.0

# File Handling
Pillow==10.1.0

# PDF Generation (for future reports)
reportlab==4.0.7
weasyprint==60.2

# Excel/CSV Export
openpyxl==3.1.2
pandas==2.1.4

# Email (for future notifications)
Flask-Mail==0.9.1

# Caching (for future optimization)
Flask-Caching==2.1.0
redis==5.0.1

# API Documentation (for future)
flask-restx==1.3.0
flasgger==0.9.7.1

# Background Tasks (for future)
celery==5.3.4

# Monitoring & Health Checks
psutil==5.9.6

# Timezone handling
babel==2.13.1


google-generativeai==0.3.1
openai
stripe
mercadopago
paypalrestsdk


Requests==2.31.0
ReportLab==4.0.7
Pillow==10.1.0
